REPL launch config now uses Couterclockwise.
The Maven based REPL console is not usable from within Eclipse as it
provides no tab-completion, no history, etc. The REPL provided by
Couterclockwise has all those features and is used together with Eclipse
anyway.

With this changes a new Counterclockwise based REPL will be started and
it will be pre-populated with whatever is defined in bootstrap.clj.

Signed-off-by: Sebastian Hoß <mail@shoss.de>